+++ src/sickle_ext/builder_ext.rs
@@ -131,7 +131,7 @@ impl scene_traits::SceneNodeBuilder for UiBuilder<'_, UiRoot>
 
     fn initialize_scene_node(ec: &mut EntityCommands)
     {
-        ec.try_insert(Node::default());
+        ec.apply(FlexNode::default());
     }
 
     fn scene_node_builder<'a>(commands: &'a mut Commands, entity: Entity) -> Self::Builder<'a>
@@ -163,7 +163,7 @@ impl scene_traits::SceneNodeBuilder for UiBuilder<'_, Entity>
 
     fn initialize_scene_node(ec: &mut EntityCommands)
     {
-        ec.try_insert(Node::default());
+        ec.apply(FlexNode::default());
     }
 
     fn scene_node_builder<'a>(commands: &'a mut Commands, entity: Entity) -> Self::Builder<'a>