+++ src/ui_bevy/ui_ext/mod.rs
@@ -5,6 +5,7 @@ mod node_wrappers;
 mod opacity;
 mod plugin;
 mod text;
+mod text_rendering;
 
 pub use component_wrappers::*;
 pub use image_node::*;
@@ -13,3 +14,4 @@ pub use node_wrappers::*;
 pub use opacity::*;
 pub use plugin::*;
 pub use text::*;
+pub(self) use text_rendering::*;